% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteTranscript.R
\name{deleteTranscript}
\alias{deleteTranscript}
\title{Delete a transcript from the corpus}
\usage{
deleteTranscript(labbcat.url, id)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}

\item{id}{The ID transcript to delete.}
}
\value{
The ID of the deleted transcript
}
\description{
This function deletes the given transcript, and all associated files.
}
\details{
For this function to work, the credentials used to connect to the server must have at
least 'edit' access.
}
\examples{
\dontrun{
## delete a transcript from the server
deleteTranscript(labbcat.url, "my-transcript.eaf")
}
}
\keyword{management}
\keyword{transcript}
