% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDictionaries.R
\name{getDictionaries}
\alias{getDictionaries}
\title{List the dictionaries available}
\usage{
getDictionaries(labbcat.url)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}
}
\value{
A named list of layer manager IDs, each of which containing a list of
dictionaries that the layer manager makes available.
}
\description{
List the dictionaries available
}
\examples{
\dontrun{
## List the dictionaries available
dictionaries <- getDictionaries("https://labbcat.canterbury.ac.nz/demo/")
}

}
\seealso{
Other dictionary functions: 
\code{\link{addDictionaryEntry}()},
\code{\link{addLayerDictionaryEntry}()},
\code{\link{deleteLexicon}()},
\code{\link{getDictionaryEntries}()},
\code{\link{loadLexicon}()},
\code{\link{removeDictionaryEntry}()},
\code{\link{removeLayerDictionaryEntry}()}
}
\concept{dictionary functions}
\keyword{dictionary}
