% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzilbb.labbcat.R
\name{getLayerIds}
\alias{getLayerIds}
\title{Gets a list of layer IDs.}
\usage{
getLayerIds(labbcat.url)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}
}
\value{
A list of layer IDs
}
\description{
Layer IDs are annotation 'types'.
}
\examples{
\dontrun{
## Connect to LaBB-CAT
labbcat.url <- "https://labbcat.canterbury.ac.nz/demo/"
labbcatCredentials(labbcat.url, "demo", "demo")

## Get names of all layers
layer.ids <- getLayerIds(labbcat.url)
}

}
\keyword{layer}
