% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzilbb.labbcat.R
\name{getParticipantIds}
\alias{getParticipantIds}
\title{Gets a list of participant IDs.}
\usage{
getParticipantIds(labbcat.url)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}
}
\value{
A list of participant IDs
}
\description{
Returns a list of participant IDs.
}
\examples{
\dontrun{
## define the LaBB-CAT URL
labbcat.url <- "https://labbcat.canterbury.ac.nz/demo/"

## specify the username/password in the script
## (only use labbcatCredentials for scripts that must execute unsupervised!)
labbcatCredentials(labbcat.url, "demo", "demo")

## List all speakers
speakers <- getParticipantIds(labbcat.url)
}

}
\keyword{participant}
\keyword{speaker}
