% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSystemAttribute.R
\name{getSystemAttribute}
\alias{getSystemAttribute}
\title{Gets the value of the given system attribute}
\usage{
getSystemAttribute(labbcat.url, attribute)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}

\item{attribute}{Name of the attribute.}
}
\value{
The value of the given attribute.
}
\description{
Gets the value of the given system attribute
}
\examples{
\dontrun{
## Get the name of the LaBB-CAT instance
title <- getSystemAttribute("https://labbcat.canterbury.ac.nz/demo/", "title")
}

}
\seealso{
\link{getLayers}
}
