% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTranscriptIds.R
\name{getTranscriptIds}
\alias{getTranscriptIds}
\title{Gets a list of transcript IDs.}
\usage{
getTranscriptIds(labbcat.url)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}
}
\value{
A list of transcript IDs
}
\description{
Returns a list of transcript IDs (i.e. transcript names).
}
\examples{
\dontrun{
## List all transcripts
transcripts <- getTranscriptIds("https://labbcat.canterbury.ac.nz/demo/")
}

}
\keyword{transcript}
