% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTranscriptIdsInCorpus.R
\name{getTranscriptIdsInCorpus}
\alias{getTranscriptIdsInCorpus}
\title{Gets a list of transcript in a corpus.}
\usage{
getTranscriptIdsInCorpus(labbcat.url, id)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}

\item{id}{The ID (name) of the corpus}
}
\value{
A list of transcript IDs
}
\description{
Returns a list of transcript IDs in the given corpus.
}
\examples{
\dontrun{
## List transcripts in the QB corpus
transcripts <- getTranscriptIdsInCorpus("https://labbcat.canterbury.ac.nz/demo/", "QB")
}

}
\keyword{corpora}
\keyword{corpus}
