% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserInfo.R
\name{getUserInfo}
\alias{getUserInfo}
\title{Gets information about the current user.}
\usage{
getUserInfo(labbcat.url)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}
}
\value{
A named list containing information about current the LaBB-CAT user.
}
\description{
Returns information about the current user, including the roles or groups they are in.
}
\examples{
\dontrun{
## List file export formats supported
me <- getUserInfo("https://labbcat.canterbury.ac.nz/demo/")

## am I an administrator?
admin <- "admin" \%in\% me$roles
}

}
\seealso{
\link{labbcatCredentials}
}
\keyword{connect}
\keyword{username}
