% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzilbb.labbcat.R
\name{labbcat.getGraphIds}
\alias{labbcat.getGraphIds}
\title{Gets a list of graph IDs.}
\usage{
labbcat.getGraphIds(labbcat)
}
\arguments{
\item{labbcat}{A LaBB-CAT instance object previously created by a call to labbcat.instance}
}
\value{
A list of graph IDs
}
\description{
Returns a list of graph IDs (i.e. transcript names).
}
\examples{
## Connect to LaBB-CAT
labbcat <- labbcat.instance("https://labbcat.canterbury.ac.nz/demo/", "demo", "demo")

## List all transcripts
transcripts <- labbcat.getGraphIds(labbcat)

}
\keyword{graph}
\keyword{transcript}
