% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzilbb.labbcat.R
\name{labbcat.getGraphIdsInCorpus}
\alias{labbcat.getGraphIdsInCorpus}
\title{Gets a list of corpus IDs.}
\usage{
labbcat.getGraphIdsInCorpus(labbcat, id)
}
\arguments{
\item{labbcat}{A LaBB-CAT instance object previously created by a call to labbcat.instance}

\item{id}{The ID (name) of the corpus}
}
\value{
A list of corpus IDs
}
\description{
Returns a list of corpora in the given 'LaBB-CAT' instance.
}
\examples{
## Connect to LaBB-CAT
labbcat <- labbcat.instance("https://labbcat.canterbury.ac.nz/demo/", "demo", "demo")

## List corpora
corpora <- labbcat.getCorpusIds(labbcat)

}
\keyword{corpora}
\keyword{corpus}
