% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzilbb.labbcat.R
\name{labbcat.getMediaTracks}
\alias{labbcat.getMediaTracks}
\title{List the predefined media tracks available for transcripts.}
\usage{
labbcat.getMediaTracks(labbcat)
}
\arguments{
\item{labbcat}{A LaBB-CAT instance object previously created by a call to labbcat.instance}
}
\value{
A list of media track definitions.
}
\description{
List the predefined media tracks available for transcripts.
}
\examples{
## Connect to LaBB-CAT
labbcat <- labbcat.instance("https://labbcat.canterbury.ac.nz/demo/", "demo", "demo")

## Get the media tracks configured in LaBB-CAT
tracks <- labbcat.getMediaTracks(labbcat)

}
\keyword{media}
\keyword{sound}
