% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzilbb.labbcat.R
\name{labbcat.getParticipantIds}
\alias{labbcat.getParticipantIds}
\title{Gets a list of participant IDs.}
\usage{
labbcat.getParticipantIds(labbcat)
}
\arguments{
\item{labbcat}{A LaBB-CAT instance object previously created by a call to labbcat.instance}
}
\value{
A list of participant IDs
}
\description{
Returns a list of participant IDs.
}
\examples{
## Connect to LaBB-CAT
labbcat <- labbcat.instance("https://labbcat.canterbury.ac.nz/demo/", "demo", "demo")

## List all speakers
speakers <- labbcat.getParticipantIds(labbcat)

}
\keyword{participant}
\keyword{speaker}
