% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labbcatTimeout.R
\name{labbcatTimeout}
\alias{labbcatTimeout}
\title{Sets the timeout for request to the LaBB-CAT server in future function calls. The
default timeout is 10 seconds.}
\usage{
labbcatTimeout(seconds = NULL)
}
\arguments{
\item{seconds}{The number of seconds before requests return with a timeout error.}
}
\value{
The request timeout in seconds
}
\description{
Sets the timeout for request to the LaBB-CAT server in future function calls. The
default timeout is 10 seconds.
}
\examples{
\dontrun{
## the request timeout
labbcatTimeout(30)
}

}
\keyword{connect}
\keyword{password}
\keyword{timeout}
\keyword{username}
