% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzilbb.labbcat.R
\name{getCorpusIds}
\alias{getCorpusIds}
\title{Gets a list of corpus IDs.}
\usage{
getCorpusIds(labbcat.url)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}
}
\value{
A list of corpus IDs
}
\description{
Returns a list of corpora in the given 'LaBB-CAT' instance.
}
\examples{
\dontrun{
## List corpora
corpora <- getCorpusIds("https://labbcat.canterbury.ac.nz/demo/")
}

}
\keyword{corpora}
