% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphIds.R
\name{getGraphIds}
\alias{getGraphIds}
\title{Gets a list of graph IDs.}
\usage{
getGraphIds(labbcat.url)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}
}
\value{
A list of graph IDs
}
\description{
Returns a list of graph IDs (i.e. transcript names).
}
\examples{
\dontrun{
## List all transcripts
transcripts <- getGraphIds("https://labbcat.canterbury.ac.nz/demo/")
}

}
\keyword{graph}
\keyword{transcript}
