% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzilbb.labbcat.R
\name{getGraphIdsInCorpus}
\alias{getGraphIdsInCorpus}
\title{Gets a list of corpus IDs.}
\usage{
getGraphIdsInCorpus(labbcat.url, id)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}

\item{id}{The ID (name) of the corpus}
}
\value{
A list of corpus IDs
}
\description{
Returns a list of corpora in the given 'LaBB-CAT' instance.
}
\examples{
\dontrun{
## define the LaBB-CAT URL
labbcat.url <- "https://labbcat.canterbury.ac.nz/demo/"

## specify the username/password in the script
## (only use labbcatCredentials for scripts that must execute unsupervised!)
labbcatCredentials(labbcat.url, "demo", "demo")

## List corpora
corpora <- getGraphIdsInCorpus(labbcat.url, "QB")
}

}
\keyword{corpora}
\keyword{corpus}
