% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getId.R
\name{getId}
\alias{getId}
\title{Gets the store's ID}
\usage{
getId(labbcat.url)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}
}
\value{
The annotation store's ID
}
\description{
The store's ID - i.e. the ID of the 'LaBB-CAT' instance.
}
\examples{
\dontrun{
## Get ID of LaBB-CAT instance
instance.id <- getId("https://labbcat.canterbury.ac.nz/demo/")
}

}
