% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayerIds.R
\name{getLayerIds}
\alias{getLayerIds}
\title{Gets a list of layer IDs.}
\usage{
getLayerIds(labbcat.url)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}
}
\value{
A list of layer IDs
}
\description{
Layer IDs are annotation 'types'.
}
\examples{
\dontrun{
## Get names of all layers
layer.ids <- getLayerIds("https://labbcat.canterbury.ac.nz/demo/")
}

}
\keyword{layer}
