% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchAlignments.R
\name{getMatchAlignments}
\alias{getMatchAlignments}
\title{Gets temporal alignments of matches on a given layer.}
\usage{
getMatchAlignments(
  labbcat.url,
  match.ids,
  layer.ids,
  target.offset = 0,
  annotations.per.layer = 1,
  anchor.confidence.min = 50,
  include.match.ids = FALSE,
  page.length = 1000,
  no.progress = FALSE
)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}

\item{match.ids}{A vector of annotation IDs, e.g. the MatchId column, or the URL column,
of a results set.}

\item{layer.ids}{A vector of layer IDs.}

\item{target.offset}{The distance from the original target of the match, e.g.
\itemize{
 \item{\emph{0} -- find annotations of the match target itself},
 \item{\emph{1} -- find annotations of the token immediately \emph{after} match target}
 \item{\emph{-1} -- find annotations of the token immediately \emph{before} match target}
}}

\item{annotations.per.layer}{The number of annotations on the given layer to
retrieve. In most cases, there's only one annotation available. However, tokens
may, for example, be annotated with `all possible phonemic transcriptions', in which
case using a value of greater than 1 for this parameter provides other phonemic
transcriptions, for tokens that have more than one.}

\item{anchor.confidence.min}{The minimum confidence for alignments, e.g.
\itemize{
 \item{\emph{0} -- return all alignments, regardless of confidence;}
 \item{\emph{50} -- return only alignments that have been at least automatically aligned;}
 \item{\emph{100} -- return only manually-set alignments.}
}}

\item{include.match.ids}{Whether or not the data frame returned includes the original
MatchId column or not.}

\item{page.length}{In order to prevent timeouts when there are a large number of
matches or the network connection is slow, rather than retrieving matches in one
big request, they are retrieved using many smaller requests. This parameter
controls the number of results retrieved per request.}

\item{no.progress}{TRUE to supress visual progress bar. Otherwise, progress bar will be
shown when interactive().}
}
\value{
A data frame with label, start time, and end time, for each layer.
}
\description{
Gets labels and start/end offsets of annotations on a given layer, identified by given
match IDs.
}
\details{
You can specify a threshold for confidence in the alignment, which is a value from 0
(not aligned) to 100 (manually aligned). The default is 50 (automatically aligned), so
only alignments that have been at least automatically aligned are specified. For cases
where there's a token but its alignment confidence falls below the threshold, a label
is returned, but the start/end times are NA.
}
\examples{
\dontrun{
## define the LaBB-CAT URL
labbcat.url <- "https://labbcat.canterbury.ac.nz/demo/"

## Perform a search
results <- getMatches(labbcat.url, list(segment="I"))

## Get the segment following the token, with alignment if it's been manually aligned
following.segment <- getMatchAlignments(labbcat.url, results$MatchId, "segment",
    target.offset=1, anchor.confidence.min=100)
}

}
\seealso{
\code{\link{getMatches}}
\code{\link{getMatchLabels}}
}
\keyword{annotation}
\keyword{label}
\keyword{layer}
