% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMediaTracks.R
\name{getMediaTracks}
\alias{getMediaTracks}
\title{List the predefined media tracks available for transcripts}
\usage{
getMediaTracks(labbcat.url)
}
\arguments{
\item{labbcat.url}{URL to the LaBB-CAT instance}
}
\value{
A list of media track definitions.
}
\description{
List the predefined media tracks available for transcripts
}
\examples{
\dontrun{
## Get the media tracks configured in LaBB-CAT
tracks <- getMediaTracks("https://labbcat.canterbury.ac.nz/demo/")
}

}
\keyword{media}
\keyword{sound}
